Refactor (#6425)

+++ ct/overseerr.sh
@@ -27,20 +27,35 @@ function update_script() {
     msg_error "No ${APP} Installation Found!"
     exit
   fi
-  msg_info "Updating $APP"
-  systemctl stop overseerr
-  cd /opt/overseerr
-  output=$(git pull)
-  $STD git pull
-  if echo "$output" | grep -q "Already up to date."; then
-    msg_ok " $APP is already up to date."
+  
+  RELEASE=$(curl -fsSL https://api.github.com/repos/sct/overseerr/releases/latest | grep "tag_name" | awk '{print substr($2, 3, length($2)-4) }')
+  if [[ ! -f ~/.overseerr ]] || [[ "${RELEASE}" != "$(cat ~/.overseerr)" ]]; then
+    msg_info "Stopping ${APP} service"
+    systemctl stop overseerr
+    msg_ok "Service stopped"
+
+    msg_info "Creating backup"
+    mv /opt/overseerr/config /opt/config_backup
+    msg_ok "Backup created"
+
+    fetch_and_deploy_gh_release "overseerr" "sct/overseerr" "tarball"
+    rm -rf /opt/overseerr/config
+
+    msg_info "Configuring ${APP} (Patience)"
+    cd /opt/overseerr
+    $STD yarn install
+    $STD yarn build
+    mv /opt/config_backup /opt/overseerr/config
+    msg_ok "Configured ${APP}"
+
+    msg_info "Starting ${APP} service"
     systemctl start overseerr
-    exit
+    msg_ok "Started ${APP} service"
+
+    msg_ok "Updated successfully!"
+  else
+    msg_ok "No update required. ${APP} is already at ${RELEASE}"
   fi
-  $STD yarn install
-  $STD yarn build
-  systemctl start overseerr
-  msg_ok "Updated $APP"
   exit
 }
